udpated middleware

+++ part4/bloglist_backend/controllers/blogs.js
@@ -1,6 +1,7 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 const User = require('../models/user')
+const jwt = require('jsonwebtoken')
 
 blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog.find({}).find({}).populate('user', 'username name')
@@ -11,19 +12,25 @@ blogsRouter.get('/', async (request, response) => {
 
 blogsRouter.post('/', async (request, response) => {
   const blog = new Blog(request.body)
-  if (request.body.user) {
-    const user = await User.findById(request.body.user.id)
+  
+  // Verify token
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
   }
+  const userObj = await User.findById(decodedToken.id)
+  const user = request.user   // This uses the middleware to get the id for user
 
-  if (request.body.user) {
-    user = await User.findById(request.body.user.id);
-  }
+  // Make blog's userfield, the current user.
+  blog.user = user
 
+  // Save blog
   const savedBlog = await blog.save()
 
-  if (user) {
-    user.blogs = user.blogs.concat(savedBlog._id)
-    await user.save()
+  // Update user's blog list
+  if (userObj) {
+    userObj.blogs = userObj.blogs.concat(savedBlog._id)
+    await userObj.save()
   }
 
   response.status(201).json(savedBlog)
@@ -50,8 +57,13 @@ blogsRouter.put('/:id', async (request, response) => {
 
 blogsRouter.delete('/:id', async (request, response) => {
   const blogId = request.params.id
-
-  await Blog.findByIdAndDelete(blogId)
+  const blog = await Blog.findById(blogId)
+  
+  //const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  //if (blog.user.toString() === decodedToken.id) {
+  if (blog.user.toString() === request.user.toString()) {
+    await Blog.findByIdAndDelete(blogId)
+  }
   response.status(204).end()
 })
